moved validators logic from node-api backend to handlers
