+++ node-api/handlers/beacon/backend.go
@@ -35,7 +35,6 @@ type Backend interface {
 	BlockBackend
 	RandaoBackend
 	StateBackend
-	ValidatorBackend
 	WithdrawalBackend
 	// GetSlotByBlockRoot retrieves the slot by a given root from the store.
 	GetSlotByBlockRoot(root common.Root) (math.Slot, error)
@@ -69,11 +68,4 @@ type StateBackend interface {
 
 type WithdrawalBackend interface {
 	PendingPartialWithdrawalsAtState(*statedb.StateDB) ([]*types.PendingPartialWithdrawalData, error)
-}
-
-type ValidatorBackend interface {
-	ValidatorBalancesByIDs(
-		slot math.Slot,
-		ids []string,
-	) ([]*types.ValidatorBalanceData, error)
 }
